back/requirements.txt
fastapi>=0.114.0,<0.115.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy
bcrypt>=4.0.0,<4.1.0
passlib[bcrypt]
pymysql
pyjwt
pytest
httpx